[APT-9577] Add DRM for downloads for offline usage : handle DRM for MPEG-Dash audio playback
Create new helper DrmMediaSourceHelper to generate the correct media item with DRM info depending on if the content is downloaded/being downloaded.

If the content is streaming, we only need to include the general DRM info so the DRM license can be fetched from the server to decrypt the encrypted content.
If the content is a download, we need to include all DRM info as well as the DRM key ID so the local DRM license can be used for decryption instead.