[APT-9578] Improve Exception Handling
ArmadilloException needs to support `cause` and `message` in the way `Exception` and most programmers expect them to be used.

Avoids initializing new `Exception()` wherever possible to prevent obfuscating the origin of an error. Never uses `object` for Exceptions, as each carry specific information on their stacktrace.

Exception messages need to be written in plain English, with indication of what component is failing when it fails. And ideally, with a hint on how to fix these errors.