[APT-10345] Update Tests
Attempts to treat downloaded content license renewals similarly to how they are treated when streamed: attempt at renewing the license when playback starts.

Also moves the main thread observer to the point in ArmadilloPlayerChoreographer where it is returned, rather than using it in the Reducer. This addresses two problems: the state being inconsistent about which thread it was on, and preventing main thread ANRs.