[APT-10393] Support Rotating Content URLs
Change the internal download tracking to be based on the ID passed in from the client, not by the URL of the audio content.  This gives the client full control over the content downloaded by Armadillo, regardless if the URL changes or not at a later time.

Beforehand, a changed URL could fail to be found in Armadillo, even though it has been downloaded. These downloads should no longer be lost.

Removed unused storage reference in the WidevineSessionEventListener.