[APT-10727] Fix seek handling
Previously, the progress position was not updating after a seek until the next playback poll. This caused problems with consumers attempting to track listening because it was not possible to accurately detect a seek. This now updates the position when the seek finishes, rather than waiting for the next poll for the position to catch up.

Also updates Reducer tests to ensure that we are chaining events properly.